Remove W3C API key
no longer required (not even recommended)
