+++ .github/workflows/lint.yml
@@ -24,18 +24,16 @@ jobs:
       - name: Install dependencies
         run: npm ci
 
-      - name: Test (including W3C API tests)
+      - name: Test (including GitHub API tests)
         env:
           GH_TOKEN: ${{ secrets.GITHUB_TOKEN }}
-          W3C_API_KEY: ${{ secrets.W3C_API_KEY }}
-        if: ${{ env.W3C_API_KEY }}
+        if: ${{ env.GH_TOKEN }}
         run: |
           npm run test
 
-      - name: Test (without W3C API tests)
+      - name: Test (without GitHub API tests)
         env:
           GH_TOKEN: ${{ secrets.GITHUB_TOKEN }}
-          W3C_API_KEY: ${{ secrets.W3C_API_KEY }}
         if: ${{ !env.W3C_API_KEY }}
         run: npm run test-pr
