+++ README.md
@@ -671,8 +671,7 @@ npm run build
 
 **Important:** The generation process will try to retrieve information about W3C
 specification from the W3C API. For that to work, the code requires the presence
-of a `config.json` file in the root folder with a `w3cApiKey` field set to a
-valid [W3C API key](https://w3c.github.io/w3c-api/) and a `GH_TOKEN` field
+of a `config.json` file in the root folder with a `GH_TOKEN` field
 set to a valid [GitHub Personal Token](https://github.com/settings/tokens)
 (default read permissions are enough).
