+++ test/fetch-groups.js
@@ -77,4 +77,68 @@ describe("fetch-groups module (without API keys)", function () {
     assert.equal(res[0].organization, "W3C");
     assert.deepStrictEqual(res[0].groups, spec.groups);
   });
+
+  describe("fetch from W3C API", () => {
+    it("handles /TR URLs", async () => {
+      const res = await fetchGroupsFor("https://www.w3.org/TR/gamepad/");
+      assert.equal(res.organization, "W3C");
+      assert.deepStrictEqual(res.groups, [{
+        name: "Web Applications Working Group",
+        url: "https://www.w3.org/groups/wg/webapps/"
+      }]);
+    });
+
+    it("handles multiple /TR URLs", async () => {
+      const specs = [
+        { url: "https://www.w3.org/TR/gamepad/" },
+        { url: "https://www.w3.org/TR/accname-1.2/" }
+      ];
+      const res = await fetchGroups(specs);
+      assert.equal(res[0].organization, "W3C");
+      assert.deepStrictEqual(res[0].groups, [{
+        name: "Web Applications Working Group",
+        url: "https://www.w3.org/groups/wg/webapps/"
+      }]);
+      assert.equal(res[1].organization, "W3C");
+      assert.deepStrictEqual(res[1].groups, [{
+        name: "Accessible Rich Internet Applications Working Group",
+        url: "https://www.w3.org/WAI/ARIA/"
+      }]);
+    });
+    it("handles SVG URLs", async () => {
+      const res = await fetchGroupsFor("https://svgwg.org/specs/animations/");
+      assert.equal(res.organization, "W3C");
+      assert.deepStrictEqual(res.groups, [{
+        name: "SVG Working Group",
+        url: "https://www.w3.org/Graphics/SVG/WG/"
+      }]);
+    });
+
+    it("handles CSS WG URLs", async () => {
+      const res = await fetchGroupsFor("https://drafts.csswg.org/css-animations-2/");
+      assert.equal(res.organization, "W3C");
+      assert.deepStrictEqual(res.groups, [{
+        name: "Cascading Style Sheets (CSS) Working Group",
+        url: "https://www.w3.org/Style/CSS/"
+      }]);
+    });
+
+    it("handles CSS Houdini TF URLs", async () => {
+      const res = await fetchGroupsFor("https://drafts.css-houdini.org/css-typed-om-2/");
+      assert.equal(res.organization, "W3C");
+      assert.deepStrictEqual(res.groups, [{
+        name: "Cascading Style Sheets (CSS) Working Group",
+        url: "https://www.w3.org/Style/CSS/"
+      }]);
+    });
+
+    it("handles CSS FXTF URLs", async () => {
+      const res = await fetchGroupsFor("https://drafts.fxtf.org/filter-effects-2/");
+      assert.equal(res.organization, "W3C");
+      assert.deepStrictEqual(res.groups, [{
+        name: "Cascading Style Sheets (CSS) Working Group",
+        url: "https://www.w3.org/Style/CSS/"
+      }]);
+    });
+  });
 });
