--- test/fetch-info-w3c.js
@@ -1,143 +1,0 @@
-/**
- * Tests for the fetch-info module that require a W3C API key
- *
- * These tests are separated from the tests that do not require a W3C API key
- * because the key cannot be exposed on pull requests from forked repositories
- * on GitHub.
- */ 
-
-const assert = require("assert");
-const fetchInfo = require("../src/fetch-info.js");
-
-const w3cApiKey = (function () {
-  try {
-    return require("../config.json").w3cApiKey;
-  }
-  catch (err) {
-    return null;
-  }
-})() ?? process.env.W3C_API_KEY;
-
-
-describe("fetch-info module (with W3C API key)", function () {
-  // Tests need to send network requests
-  this.slow(5000);
-  this.timeout(30000);
-
-  function getW3CSpec(shortname, series) {
-    const spec = { shortname, url: `https://www.w3.org/TR/${shortname}/` };
-    if (series) {
-      spec.series = { shortname: series };
-    }
-    return spec;
-  }
-
-  describe("W3C API key", () => {
-    it("is defined otherwise tests cannot pass", () => {
-      assert.ok(w3cApiKey);
-    });
-  });
-
-
-  describe("fetch from W3C API", () => {
-    it("works on a TR spec", async () => {
-      const spec = getW3CSpec("hr-time-2", "hr-time");
-      const info = await fetchInfo([spec], { w3cApiKey });
-      assert.ok(info[spec.shortname]);
-      assert.equal(info[spec.shortname].source, "w3c");
-      assert.equal(info[spec.shortname].release.url, spec.url);
-      assert.equal(info[spec.shortname].release.status, "Recommendation");
-      assert.equal(info[spec.shortname].nightly.url, "https://w3c.github.io/hr-time/");
-      assert.equal(info[spec.shortname].nightly.status, "Editor's Draft");
-      assert.equal(info[spec.shortname].title, "High Resolution Time Level 2");
-
-      assert.ok(info.__series);
-      assert.ok(info.__series["hr-time"]);
-      assert.equal(info.__series["hr-time"].currentSpecification, "hr-time-3");
-      assert.equal(info.__series["hr-time"].title, "High Resolution Time");
-    });
-
-    it("can operate on multiple specs at once", async () => {
-      const spec = getW3CSpec("hr-time-2", "hr-time");
-      const other = getW3CSpec("tracking-dnt", "tracking-dnt");
-      const info = await fetchInfo([spec, other], { w3cApiKey });
-      assert.ok(info[spec.shortname]);
-      assert.equal(info[spec.shortname].source, "w3c");
-      assert.equal(info[spec.shortname].release.url, spec.url);
-      assert.equal(info[spec.shortname].release.status, "Recommendation");
-      assert.equal(info[spec.shortname].nightly.url, "https://w3c.github.io/hr-time/");
-      assert.equal(info[spec.shortname].nightly.status, "Editor's Draft");
-      assert.equal(info[spec.shortname].title, "High Resolution Time Level 2");
-
-      assert.ok(info[other.shortname]);
-      assert.equal(info[other.shortname].source, "w3c");
-      assert.equal(info[other.shortname].release.url, other.url);
-      assert.equal(info[other.shortname].release.status, "Discontinued Draft");
-      assert.equal(info[other.shortname].nightly.url, "https://w3c.github.io/dnt/drafts/tracking-dnt.html");
-      assert.equal(info[other.shortname].nightly.status, "Editor's Draft");
-      assert.equal(info[other.shortname].title, "Tracking Preference Expression (DNT)");
-
-      assert.ok(info.__series);
-      assert.ok(info.__series["hr-time"]);
-      assert.ok(info.__series["tracking-dnt"]);
-      assert.equal(info.__series["hr-time"].currentSpecification, "hr-time-3");
-      assert.equal(info.__series["tracking-dnt"].currentSpecification, "tracking-dnt");
-    });
-
-    it("throws when W3C API key is invalid", async () => {
-      assert.rejects(
-        fetchInfo([getW3CSpec("selectors-3")], { w3cApiKey: "invalid" }),
-        /^W3C API returned an error, status code is 403$/);
-    });
-  });
-
-
-  describe("fetch from Specref", () => {
-    it("works on a WHATWG spec", async () => {
-      const spec = {
-        url: "https://dom.spec.whatwg.org/",
-        shortname: "dom"
-      };
-      const info = await fetchInfo([spec], { w3cApiKey });
-      assert.ok(info[spec.shortname]);
-      assert.equal(info[spec.shortname].source, "specref");
-      assert.equal(info[spec.shortname].nightly.url, "https://dom.spec.whatwg.org/");
-      assert.equal(info[spec.shortname].nightly.status, "Living Standard");
-      assert.equal(info[spec.shortname].title, "DOM Standard");
-    });
-  });
-
-
-  describe("fetch from all sources", () => {
-    it("merges info from sources", async () => {
-      const w3c = getW3CSpec("presentation-api");
-      const whatwg = {
-        url: "https://html.spec.whatwg.org/multipage/",
-        shortname: "html"
-      };
-      const other = {
-        url: "https://speced.github.io/bikeshed/",
-        shortname: "bikeshed"
-      };
-      const info = await fetchInfo([w3c, whatwg, other], { w3cApiKey });
-      assert.ok(info[w3c.shortname]);
-      assert.equal(info[w3c.shortname].source, "w3c");
-      assert.equal(info[w3c.shortname].release.url, w3c.url);
-      assert.equal(info[w3c.shortname].nightly.url, "https://w3c.github.io/presentation-api/");
-      assert.equal(info[w3c.shortname].nightly.status, "Editor's Draft");
-      assert.equal(info[w3c.shortname].title, "Presentation API");
-
-      assert.ok(info[whatwg.shortname]);
-      assert.equal(info[whatwg.shortname].source, "specref");
-      assert.equal(info[whatwg.shortname].nightly.url, whatwg.url);
-      assert.equal(info[whatwg.shortname].nightly.status, "Living Standard");
-      assert.equal(info[whatwg.shortname].title, "HTML Standard");
-
-      assert.ok(info[other.shortname]);
-      assert.equal(info[other.shortname].source, "spec");
-      assert.equal(info[other.shortname].nightly.url, other.url);
-      assert.equal(info[other.shortname].nightly.status, "Living Standard");
-      assert.equal(info[other.shortname].title, "Bikeshed Documentation");      
-    });
-  });
-});
